# backend/openapi.yaml
openapi: 3.0.3
info:
  title: TriLink Platform API
  description: |
    Comprehensive API documentation for the TriLink Platform - a multi-party procurement and logistics management system.
    
    ## Authentication
    Most endpoints require JWT Bearer token authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your-access-token>
    ```
    
    ## Roles & Permissions
    The platform supports the following roles:
    - **Buyer**: Creates purchase requests, evaluates bids, manages contracts
    - **Supplier**: Responds to RFQs, submits bids
    - **Logistics**: Manages shipments, updates GPS tracking
    - **Clearance**: Handles customs clearance
    - **Service Provider**: Provides various services
    - **Government**: View-only access with dispute resolution capabilities
    - **Admin**: Full system access
    
    ## Company Isolation
    Most resources are isolated by company. Users can only access resources belonging to their company unless they have Admin role.
  version: 1.0.0
  contact:
    name: TriLink Platform Support
    email: support@trilink.com
  license:
    name: Proprietary
servers:
  - url: http://localhost:3000/api
    description: Development server
  - url: https://api.trilink.com/api
    description: Production server
...
paths:
...
  /purchase-requests:
    post:
      tags:
        - Purchase Requests
      summary: Create purchase request
      security:
        - bearerAuth: []
      description: Creates a new purchase request. Buyer role only.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePurchaseRequestRequest'
      responses:
        '201':
          description: Purchase request created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponsePurchaseRequest'

    get:
      tags:
        - Purchase Requests
      summary: Get purchase requests
      security:
        - bearerAuth: []
      description: Retrieves purchase requests for the authenticated user's company. Buyer role only.
      responses:
        '200':
          description: List of purchase requests
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/PurchaseRequestResponse'
  /purchase-requests/{id}:
    get:
      tags:
        - Purchase Requests
      summary: Get purchase request by ID
      security:
        - bearerAuth: []
      description: Retrieves a specific purchase request. Buyer role only.
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '200':
          description: Purchase request details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponsePurchaseRequest'

    patch:
      tags:
        - Purchase Requests
      summary: Update purchase request
      security:
        - bearerAuth: []
      description: Updates a purchase request. Buyer role only.
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePurchaseRequestRequest'
      responses:
        '200':
          description: Purchase request updated successfully

    delete:
      tags:
        - Purchase Requests
      summary: Delete purchase request
      security:
        - bearerAuth: []
      description: Deletes a purchase request. Buyer role only.
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '200':
          description: Purchase request deleted successfully
...
  /purchase-requests/{id}/approve:
    post:
      tags:
        - Purchase Requests
      summary: Approve purchase request
      security:
        - bearerAuth: []
      description: Approves a purchase request
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '200':
          description: Purchase request approved successfully
...
  /rfqs/purchase-request/{purchaseRequestId}:
    get:
      tags:
        - RFQs
      summary: Get RFQs by purchase request
      security:
        - bearerAuth: []
      description: Retrieves all RFQs for a specific purchase request
      parameters:
        - name: purchaseRequestId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of RFQs
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/RFQResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Bearer token authentication
  parameters:
...
    ResourceId:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: Resource ID
  schemas:
...
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
        error:
          type: string
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
        requestId:
          type: string
...
    Coordinates:
      type: object
      properties:
        lat:
          type: number
          format: double
        lng:
          type: number
          format: double
...
    Address:
      type: object
      properties:
        street:
          type: string
        city:
          type: string
        state:
          type: string
        country:
          type: string
        zipCode:
          type: string
    DeliveryLocation:
      allOf:
        - $ref: '#/components/schemas/Address'
        - type: object
          properties:
            coordinates:
              $ref: '#/components/schemas/Coordinates'
...
    PurchaseRequestItem:
      type: object
      properties:
        name:
          type: string
        quantity:
          type: number
        unit:
          type: string
        specifications:
          type: string
        estimatedPrice:
          type: number
    CreatePurchaseRequestRequest:
      type: object
      required:
        - title
        - description
        - items
        - budget
        - deliveryLocation
        - requiredDeliveryDate
      properties:
        title:
          type: string
        description:
          type: string
        items:
          type: array
          items:
            $ref: '#/components/schemas/PurchaseRequestItem'
        budget:
          type: number
        currency:
          type: string
          default: AED
        deliveryLocation:
          $ref: '#/components/schemas/DeliveryLocation'
        requiredDeliveryDate:
          type: string
          format: date-time
    UpdatePurchaseRequestRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        items:
          type: array
          items:
            $ref: '#/components/schemas/PurchaseRequestItem'
        budget:
          type: number
        currency:
          type: string
        deliveryLocation:
          $ref: '#/components/schemas/DeliveryLocation'
        requiredDeliveryDate:
          type: string
          format: date-time
        status:
          type: string
          enum: [draft, submitted]
    PurchaseRequestResponse:
      type: object
      properties:
        _id:
          type: string
        buyerId:
          type: string
        companyId:
          type: string
        title:
          type: string
        description:
          type: string
        items:
          type: array
          items:
            $ref: '#/components/schemas/PurchaseRequestItem'
        budget:
          type: number
        currency:
          type: string
        deliveryLocation:
          $ref: '#/components/schemas/DeliveryLocation'
        requiredDeliveryDate:
          type: string
          format: date-time
        status:
          type: string
        rfqGenerated:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ApiResponsePurchaseRequest:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/PurchaseRequestResponse'

    # RFQ schemas
    RFQItem:
      type: object
      properties:
        name:
          type: string
        quantity:
          type: number
        unit:
          type: string
        specifications:
          type: string
...
    RFQResponse:
      type: object
      properties:
        _id:
          type: string
        rfqNumber:
          type: string
        purchaseRequestId:
          type: string
        companyId:
          type: string
        type:
          type: string
        targetRole:
          type: string
        targetCompanyType:
          type: string
        title:
          type: string
        description:
          type: string
        items:
          type: array
          items:
            $ref: '#/components/schemas/RFQItem'
        budget:
          type: number
        currency:
          type: string
        deliveryLocation:
          $ref: '#/components/schemas/Address'
        requiredDeliveryDate:
          type: string
          format: date-time
        deadline:
          type: string
          format: date-time
        status:
          type: string
        anonymousBuyer:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...